requirements.txt
# TTS Gateway 依赖

# Web 框架
fastapi>=0.109.0
uvicorn>=0.27.0

# HTTP 客户端
httpx>=0.26.0

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 配置
PyYAML>=6.0.1

# 文件上传
python-multipart>=0.0.6
